[FIRRTL] Add layer block support to AddSeqMemPorts
Update the AddSeqMemPorts pass to support layers.  Namely, this will now
allow for input ports to be added to memories declared under layer block.
If a user tries to add output ports, then produce an error.

Generally, this entire pass likely doesn't make sense for layer blocks at
all.  I.e., this pass is entirely about adding MBIST ports to memories.
Memories that are declared under layerblocks are likely only for
verification use and adding MBIST ports to these doesn't really make
sense.  Nonetheless, it is reasonable to have this pass do something sane
if it encounters a user request like this.

This made a small change to the pass output where the connects created are
placed immediately following the memory instance.  This is necessary to
create the connects inside the layer blocks.  However, it is also slightly
better output than before.  (This changed one test superficially.)

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>